# Update benchmark to use Value