--- a/packages/slate-html-serializer/benchmark/html-serializer/serialize.js
+++ b/packages/slate-html-serializer/benchmark/html-serializer/serialize.js
@@ -35,7 +35,7 @@ export default function (state) {
 }
 
 export const input = (
-  <state>
+  <value>
     <document>
       {Array.from(Array(10)).map(() => (
         <quote>
@@ -45,5 +45,5 @@ export const input = (
         </quote>
       ))}
     </document>
-  </state>
+  </value>
 )