--- a/packages/slate/benchmark/models/get-inlines.js
+++ b/packages/slate/benchmark/models/get-inlines.js
@@ -3,12 +3,12 @@
 
 import h from '../../test/helpers/h'
 
-export default function (state) {
-  state.document.getInlines()
+export default function (value) {
+  value.document.getInlines()
 }
 
 export const input = (
-  <state>
+  <value>
     <document>
       {Array.from(Array(10)).map(() => (
         <quote>
@@ -20,5 +20,5 @@ export const input = (
         </quote>
       ))}
     </document>
-  </state>
+  </value>
 )