This just fixes the benchmark to be run in the latest version.

```
  benchmarks
    html-serializer
      deserialize
        8.68 → 7.20 ops/sec
      serialize
        852.30 → 775.35 ops/sec
    plain-serializer
      deserialize
        254.89 → 247.64 ops/sec
      serialize
        7248.68 → 7343.98 ops/sec
    rendering
      normal
        78.13 → 78.87 ops/sec
    changes
      delete-backward
        72691.07 → 75240.82 ops/sec
      delete-forward
        13114.26 → 15398.91 ops/sec
      insert-text-by-key-multiple
        75.88 → 85.22 ops/sec
      insert-text-by-key
        1053.04 → 962.39 ops/sec
      insert-text
        952.91 → 1014.81 ops/sec
      normalize
        3037.60 → 3518.49 ops/sec
      split-block
        59.17 → 61.41 ops/sec
    models
      from-json
        118.76 → 126.34 ops/sec
      get-blocks-at-range
        549113.52 → 564613.16 ops/sec
      get-blocks
        2852520.48 → 2724271.59 ops/sec
      get-characters-at-range
        601222.96 → 589214.31 ops/sec
      get-characters
        2837836.98 → 2705111.14 ops/sec
      get-inlines-at-range
        604865.22 → 590231.48 ops/sec
      get-inlines
        2879718.43 → 2774580.39 ops/sec
      get-leaves
        1412061.43 → 1376369.53 ops/sec
      get-marks-at-range
        593833.83 → 593306.95 ops/sec
      get-marks
        2844220.51 → 2682124.81 ops/sec
      get-path
        657133.39 → 623939.48 ops/sec
      get-texts-at-range
        620414.28 → 611390.92 ops/sec
      get-texts
        2859405.29 → 2751465.12 ops/sec
      has-node-multiple
        76126.65 → 75901.32 ops/sec
      has-node
        679765.27 → 670295.53 ops/sec
      to-json
        2183.38 → 2405.11 ops/sec
      update-node
        15044.86 → 13560.42 ops/sec

✨  Done in 60.21s.
```